+++ src/main/java/roomescape/infrastructure/auth/JwtProvider.java
@@ -1,4 +1,4 @@
-package roomescape.web.auth;
+package roomescape.infrastructure.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
